python-mqtt/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
python-dotenv==1.0.1 